Moved helper #9993
